changed resources to use signed integers for sizes

+++ base/include/shoveler/view/resources.h
@@ -10,7 +10,7 @@
 typedef struct {
 	const char *typeId;
 	const unsigned char *buffer;
-	size_t bufferSize;
+	int bufferSize;
 } ShovelerViewResourceConfiguration;
 
 static const char *shovelerViewResourceComponentName = "resource";
